requirements.txt
llm
google-generativeai
streamlit